Edited Lab manual to add a section for setting up Jupyter
After much effort, Jeff, Yuyan and I figured out how to change the version of Python and gcc used by Jupyter so that Scanpy can be installed. The versions prescribed by the HPC Wiki do not work.
The new manual basically walks through the same steps as the HPC guide, except for the addition of changing the python and gcc to the right version. This manual is only for Mac clients but so far as I can tell, the same process applies to Windows clients, for which they should refer to the HPC Wiki.